Pass image data directly to `ZlibStream`, bypassing `ChunkState::raw_bytes`.
Before these changes we would store all image data in
`ChunkState::raw_bytes` before attempting to decompress them via
`ZlibStream`.  This is undesirable for performance, because 1) it copies
data unnecessarily into and out of `raw_bytes` and 2) by the time we
have stored all the image data, the start of it is long gone from the L1
cache.

After these changes, the image data is passed directly into
`ZlibStream`.